+++ src/connection_settings.h
@@ -37,6 +37,7 @@ struct ConnectionSettings {
   bool fullscreen = false;
   std::string serial_device = "";
   unsigned int serial_rate = 9600;
+  bool insecure = false;
 
   std::string sora_signaling_host = "wss://example.com/signaling";
   std::string sora_channel_id;