+++ src/rtc/manager.cpp
@@ -17,6 +17,7 @@
 #include "modules/video_capture/video_capture_factory.h"
 #include "observer.h"
 #include "rtc_base/logging.h"
+#include "rtc_base/openssl_certificate.h"
 #include "rtc_base/ssl_adapter.h"
 #include "scalable_track_source.h"
 #include "util.h"
@@ -32,14 +33,16 @@
 #include "ros/ros_audio_device_module.h"
 #endif
 
-#if USE_MMAL_ENCODER || USE_JETSON_ENCODER
+#if USE_MMAL_ENCODER || USE_JETSON_ENCODER || USE_NVCODEC_ENCODER
 #include "api/video_codecs/video_encoder_factory.h"
 #include "hw_video_encoder_factory.h"
 #endif
-#if USE_JETSON_ENCODER
+#if USE_MMAL_ENCODER || USE_JETSON_ENCODER
 #include "api/video_codecs/video_decoder_factory.h"
 #include "hw_video_decoder_factory.h"
 #endif
+
+#include "ssl_verifier.h"
 
 RTCManager::RTCManager(
     ConnectionSettings conn_settings,
@@ -96,7 +99,7 @@ RTCManager::RTCManager(
   media_dependencies.video_encoder_factory = CreateObjCEncoderFactory();
   media_dependencies.video_decoder_factory = CreateObjCDecoderFactory();
 #else
-#if USE_MMAL_ENCODER || USE_JETSON_ENCODER
+#if USE_MMAL_ENCODER || USE_JETSON_ENCODER || USE_NVCODEC_ENCODER
   media_dependencies.video_encoder_factory =
       std::unique_ptr<webrtc::VideoEncoderFactory>(
           absl::make_unique<HWVideoEncoderFactory>());
@@ -104,7 +107,7 @@ RTCManager::RTCManager(
   media_dependencies.video_encoder_factory =
       webrtc::CreateBuiltinVideoEncoderFactory();
 #endif
-#if USE_JETSON_ENCODER
+#if USE_MMAL_ENCODER || USE_JETSON_ENCODER
   media_dependencies.video_decoder_factory =
       std::unique_ptr<webrtc::VideoDecoderFactory>(
           absl::make_unique<HWVideoDecoderFactory>());
@@ -191,6 +194,21 @@ void RTCManager::SetDataManager(RTCDataManager* data_manager) {
   _data_manager = data_manager;
 }
 
+class RTCSSLVerifier : public rtc::SSLCertificateVerifier {
+  bool insecure_;
+
+ public:
+  RTCSSLVerifier(bool insecure) : insecure_(insecure) {}
+  bool Verify(const rtc::SSLCertificate& certificate) override {
+    // insecure の場合は証明書をチェックしない
+    if (insecure_) {
+      return true;
+    }
+    return SSLVerifier::VerifyX509(
+        static_cast<const rtc::OpenSSLCertificate&>(certificate).x509());
+  }
+};
+
 std::shared_ptr<RTCConnection> RTCManager::createConnection(
     webrtc::PeerConnectionInterface::RTCConfiguration rtc_config,
     RTCMessageSender* sender) {
@@ -198,9 +216,17 @@ std::shared_ptr<RTCConnection> RTCManager::createConnection(
   rtc_config.sdp_semantics = webrtc::SdpSemantics::kUnifiedPlan;
   std::unique_ptr<PeerConnectionObserver> observer(
       new PeerConnectionObserver(sender, _receiver, _data_manager));
+  webrtc::PeerConnectionDependencies dependencies(observer.get());
+
+  // WebRTC の SSL 接続の検証は自前のルート証明書(rtc_base/ssl_roots.h)でやっていて、
+  // その中に Let's Encrypt の証明書が無いため、接続先によっては接続できないことがある。
+  //
+  // それを解消するために tls_cert_verifier を設定して自前で検証を行う。
+  dependencies.tls_cert_verifier = std::unique_ptr<rtc::SSLCertificateVerifier>(
+      new RTCSSLVerifier(_conn_settings.insecure));
+
   rtc::scoped_refptr<webrtc::PeerConnectionInterface> connection =
-      _factory->CreatePeerConnection(rtc_config, nullptr, nullptr,
-                                     observer.get());
+      _factory->CreatePeerConnection(rtc_config, std::move(dependencies));
   if (!connection) {
     RTC_LOG(LS_ERROR) << __FUNCTION__ << ": CreatePeerConnection failed";
     return nullptr;