+++ src/sora/sora_websocket_client.cpp
@@ -10,6 +10,7 @@
 #include <nlohmann/json.hpp>
 
 #include "momo_version.h"
+#include "ssl_verifier.h"
 #include "url_parts.h"
 #include "util.h"
 
@@ -34,7 +35,7 @@ bool SoraWebsocketClient::parseURL(URLParts& parts) const {
 
 boost::asio::ssl::context SoraWebsocketClient::createSSLContext() const {
   boost::asio::ssl::context ctx(boost::asio::ssl::context::tlsv12);
-  ctx.set_default_verify_paths();
+  //ctx.set_default_verify_paths();
   ctx.set_options(boost::asio::ssl::context::default_workarounds |
                   boost::asio::ssl::context::no_sslv2 |
                   boost::asio::ssl::context::no_sslv3 |
@@ -74,10 +75,23 @@ void SoraWebsocketClient::reset() {
 
   if (parseURL(parts_)) {
     auto ssl_ctx = createSSLContext();
-    boost::beast::websocket::stream<
-        boost::asio::ssl::stream<boost::asio::ip::tcp::socket>>
-        wss(ioc_, ssl_ctx);
     ws_.reset(new Websocket(ioc_, std::move(ssl_ctx)));
+    ws_->nativeSecureSocket().next_layer().set_verify_mode(
+        boost::asio::ssl::verify_peer);
+    ws_->nativeSecureSocket().next_layer().set_verify_callback(
+        [insecure = conn_settings_.insecure](
+            bool preverified, boost::asio::ssl::verify_context& ctx) {
+          if (preverified) {
+            return true;
+          }
+          // insecure の場合は証明書をチェックしない
+          if (insecure) {
+            return true;
+          }
+          X509* cert = X509_STORE_CTX_get_current_cert(ctx.native_handle());
+          return SSLVerifier::VerifyX509(cert);
+        });
+
     // SNI の設定を行う
     if (!SSL_set_tlsext_host_name(
             ws_->nativeSecureSocket().next_layer().native_handle(),
@@ -262,6 +276,13 @@ void SoraWebsocketClient::doSendPong() {
   json json_message = {{"type", "pong"}};
   ws_->sendText(json_message.dump());
 }
+void SoraWebsocketClient::doSendPong(
+    const rtc::scoped_refptr<const webrtc::RTCStatsReport>& report) {
+  std::string stats = report->ToJson();
+  json json_message = {{"type", "pong"}, {"stats", stats}};
+  std::string str = R"({"type":"pong","stats":)" + stats + "}";
+  ws_->sendText(std::move(str));
+}
 
 void SoraWebsocketClient::createPeerFromConfig(json jconfig) {
   webrtc::PeerConnectionInterface::RTCConfiguration rtc_config;
@@ -357,7 +378,16 @@ void SoraWebsocketClient::onRead(boost::system::error_code ec,
       return;
     }
     watchdog_.reset();
-    doSendPong();
+    bool stats = json_message.value("stats", false);
+    if (stats) {
+      connection_->getStats(
+          [this](
+              const rtc::scoped_refptr<const webrtc::RTCStatsReport>& report) {
+            doSendPong(report);
+          });
+    } else {
+      doSendPong();
+    }
   }
 }